Separate archetype system, asset and module caches into lib/
